+++ ReleaseNotes.md
@@ -1,10 +1,19 @@
+### New in 0.0.4 (Released 2015/08/10)
+* Seperate out BindingSettings
+* Add / Remove site binding
+* Add / Remove site application
+* Change Debug to Information logs
+
 ### New in 0.0.3 (Released 2015/08/09)
-* Process model settings
-* Request tracing settings
+* ProcessModel settings
+* RequestTracing settings
 * Shared Web and Ftp create function
 
 ### New in 0.0.2 (Released 2015/08/06)
 * Rewrite by Phillip Sharpe
+* Add support for remote servers
+* Replace mixed functionality deploy methods with focused create / remove methods
+* Add missing Aliases for sites and pools
 
 ### New in 0.0.1 (Released 2015/03/31)
 * First release by Sergio Silveira.